refactor(suite): Refactor & polish  welcome screen layout
