+++ packages/suite/src/components/suite/PrerequisitesGuide/DeviceNoFirmware.tsx
@@ -1,6 +1,6 @@
 import { MouseEventHandler } from 'react';
 
-import { Button } from '@trezor/components';
+import { Banner } from '@trezor/components';
 
 import { goto } from 'src/actions/suite/routerActions';
 import { TroubleshootingTips } from 'src/components/suite';
@@ -19,10 +19,11 @@ export const DeviceNoFirmware = () => {
         <TroubleshootingTips
             label={<Translation id="TR_NO_FIRMWARE" />}
             cta={
-                <Button onClick={handleClick}>
+                <Banner.Button onClick={handleClick}>
                     <Translation id="TR_GO_TO_ONBOARDING" />
-                </Button>
+                </Banner.Button>
             }
+            variant="info"
             items={[
                 {
                     key: 'device-firmware-missing',
