+++ packages/suite/src/components/suite/PrerequisitesGuide/DeviceUnknown.tsx
@@ -4,6 +4,7 @@ import { Translation } from 'src/components/suite/Translation';
 export const DeviceUnknown = () => (
     <TroubleshootingTips
         label={<Translation id="TR_UNKNOWN_DEVICE" />}
+        variant="warning"
         items={[
             {
                 key: 'device-unknown',
