+++ packages/suite/src/components/suite/PrerequisitesGuide/MultiShareBackupInProgress.tsx
@@ -4,6 +4,7 @@ import { Translation } from 'src/components/suite/Translation';
 export const MultiShareBackupInProgress = () => (
     <TroubleshootingTips
         label={<Translation id="TR_MULTI_SHARE_BACKUP_IN_PROGRESS" />}
+        variant="info"
         items={[
             {
                 key: 'multi-share-backup-in-progress',
