+++ packages/suite/src/components/suite/PrerequisitesGuide/NoTransport.tsx
@@ -29,6 +29,7 @@ const Tips = ({ items }: { items: TroubleshootingTipsItem[] }) => (
     // Desktop app should have Bridge transport layer available as it is built-in, if it is not available we fucked up something.
     <TroubleshootingTips
         label={<Translation id="TR_TROUBLESHOOTING_DEVICE_NOT_DETECTED" />}
+        variant="warning"
         items={items}
         data-testid="@connect-device-prompt/bridge-not-running"
     />
