+++ packages/suite/src/components/suite/PrerequisitesGuide/PrerequisitesGuide.tsx
@@ -1,40 +1,22 @@
-import { JSX, useMemo } from 'react';
-
 import { motion } from 'framer-motion';
-import styled from 'styled-components';
 
+import { DEFAULT_FLAGSHIP_MODEL } from '@suite-common/suite-constants';
 import {
     deviceNeedsAttention,
+    getDeviceInternalModel,
     getStatus,
     shouldDisplayInitialWarningIcon,
 } from '@suite-common/suite-utils';
 import { selectSelectedDevice } from '@suite-common/wallet-core';
-import { Column, motionEasing } from '@trezor/components';
-import { spacings } from '@trezor/theme';
+import { Box, Column, Text, motionEasing } from '@trezor/components';
+import { DeviceWithScene } from '@trezor/product-components';
 
-import { ConnectDevicePrompt } from 'src/components/suite';
+import { getMessageId } from 'src/components/suite';
 import { useSelector } from 'src/hooks/suite';
 import { selectPrerequisite } from 'src/selectors/suite/suiteSelectors';
 
-import { DeviceAcquire } from './DeviceAcquire';
-import { DeviceBootloader } from './DeviceBootloader';
-import { DeviceConnect } from './DeviceConnect';
-import { DeviceDisconnectRequired } from './DeviceDisconnectRequired';
-import { DeviceInitialize } from './DeviceInitialize';
-import { DeviceNoFirmware } from './DeviceNoFirmware';
-import { DeviceRecoveryMode } from './DeviceRecoveryMode';
-import { DeviceSeedless } from './DeviceSeedless';
-import { DeviceTrezorHostProtocolPair } from './DeviceTrezorHostProtocolPair';
-import { DeviceUnknown } from './DeviceUnknown';
-import { DeviceUnreadable } from './DeviceUnreadable';
-import { DeviceUpdateRequired } from './DeviceUpdateRequired';
-import { DeviceUsedElsewhere } from './DeviceUsedElsewhere';
-import { MultiShareBackupInProgress } from './MultiShareBackupInProgress';
-import { NoTransport } from './NoTransport';
-
-const BottomAnimatedContainer = styled(motion.div)`
-    display: flex;
-`;
+import { Translation } from '../Translation';
+import { BannerAndTroubleshooting } from './BannerAndTroubleshooting';
 
 type PrerequisitesGuideProps = {
     showDeviceImage?: boolean;
@@ -44,71 +26,87 @@ export const PrerequisitesGuide = ({ showDeviceImage = true }: PrerequisitesGuid
     const device = useSelector(selectSelectedDevice);
     const prerequisite = useSelector(selectPrerequisite);
 
-    const TipComponent = useMemo(
-        () => (): JSX.Element => {
-            switch (prerequisite) {
-                case 'no-transport':
-                    return <NoTransport />;
-                case 'device-disconnect-required':
-                    return <DeviceDisconnectRequired />;
-                case 'device-disconnected':
-                    return <DeviceConnect />;
-                case 'device-unacquired':
-                case 'device-busy':
-                    return <DeviceAcquire />;
-                case 'device-thp-locked':
-                    return <DeviceTrezorHostProtocolPair />;
-                case 'device-used-elsewhere':
-                    return <DeviceUsedElsewhere />;
-                case 'device-unreadable':
-                    return <DeviceUnreadable />;
-                case 'device-unknown':
-                    return <DeviceUnknown />;
-                case 'device-seedless':
-                    return <DeviceSeedless />;
-                case 'device-recovery-mode':
-                    return <DeviceRecoveryMode />;
-                case 'device-initialize':
-                    return <DeviceInitialize />;
-                case 'device-bootloader':
-                    return <DeviceBootloader />;
-                case 'firmware-missing':
-                    return <DeviceNoFirmware />;
-                case 'firmware-required':
-                    return <DeviceUpdateRequired />;
-                case 'multi-share-backup-in-progress':
-                    return <MultiShareBackupInProgress />;
-                default:
-                    return <></>;
-            }
-        },
-        [prerequisite],
+    const deviceStatus = (device && getStatus(device)) ?? null;
+    const selectedDevice = useSelector(selectSelectedDevice);
+    const selectedDeviceModelInternal = selectedDevice
+        ? getDeviceInternalModel(selectedDevice)
+        : DEFAULT_FLAGSHIP_MODEL;
+
+    const showWarning =
+        !!(device && deviceStatus && deviceNeedsAttention(deviceStatus)) ||
+        prerequisite === 'no-transport';
+    const showWarningIcon = shouldDisplayInitialWarningIcon(deviceStatus);
+
+    const texts = getMessageId({
+        connected: !!device,
+        showWarning: showWarningIcon ?? showWarning,
+        deviceStatus,
+        prerequisite,
+    });
+
+    const DeviceImage = () => {
+        if (!showDeviceImage) return null;
+
+        return (
+            <Box margin={{ top: 40, bottom: 60 }}>
+                <DeviceWithScene
+                    deviceModel={selectedDeviceModelInternal}
+                    unitColor={selectedDevice?.features?.unit_color}
+                    height={300}
+                />
+            </Box>
+        );
+    };
+
+    const Heading = () => (
+        <Text typographyStyle="titleMedium" textWrap="balance" align="center">
+            <Translation id={texts.heading} />
+        </Text>
     );
 
-    const deviceStatus = (device && getStatus(device)) ?? null;
+    const Description = () => {
+        if (!texts.description) return null;
+
+        return (
+            <Text variant="tertiary" align="center" margin={{ top: 12 }}>
+                {texts.description}
+            </Text>
+        );
+    };
+
+    const TopAnimation = ({ children }: { children: React.ReactNode }) => (
+        <motion.div
+            initial={{ opacity: 0, y: -50 }}
+            animate={{ opacity: 1, y: -0 }}
+            transition={{ delay: 0.2, duration: 0.4, ease: motionEasing.enter }}
+            data-testid="@connect-device-prompt"
+        >
+            {children}
+        </motion.div>
+    );
+
+    const BottomAnimation = ({ children }: { children: React.ReactNode }) => (
+        <motion.div
+            initial={{ opacity: 0 }}
+            animate={{ opacity: 1 }}
+            transition={{ delay: 0.6, duration: 0.5, ease: motionEasing.enter }}
+        >
+            {children}
+        </motion.div>
+    );
 
     return (
-        <Column alignItems="center" gap={spacings.xxxl} margin={{ vertical: 40 }}>
-            <ConnectDevicePrompt
-                connected={!!device}
-                deviceStatus={deviceStatus}
-                showWarning={
-                    !!(device && deviceStatus && deviceNeedsAttention(deviceStatus)) ||
-                    prerequisite === 'no-transport'
-                }
-                showWarningIcon={shouldDisplayInitialWarningIcon(deviceStatus)}
-                prerequisite={prerequisite}
-                showImage={showDeviceImage}
-            />
-            <BottomAnimatedContainer
-                initial={{ opacity: 0 }}
-                animate={{ opacity: 1 }}
-                transition={{ delay: 0.6, duration: 0.5, ease: motionEasing.enter }}
-            >
-                <Column alignItems="center" justifyContent="center" gap={spacings.xxxxl}>
-                    <TipComponent />
+        <>
+            <TopAnimation>
+                <Column alignItems="center">
+                    <DeviceImage />
+                    <Heading />
+                    <Description />
                 </Column>
-            </BottomAnimatedContainer>
-        </Column>
+            </TopAnimation>
+            <BottomAnimation>
+                <BannerAndTroubleshooting prerequisite={prerequisite} />
+            </BottomAnimation>
+        </>
     );
 };
