+++ packages/suite/src/components/suite/troubleshooting/TroubleshootingTipsToggle.tsx
@@ -10,12 +10,14 @@ type TroubleshootingTipsToggleProps = {
 };
 
 export const TroubleshootingTipsToggle = ({ children }: TroubleshootingTipsToggleProps) => (
-    <Row justifyContent="center" flex="1" margin={{ bottom: spacings.xs }}>
-        <Row gap={spacings.xs} cursor="pointer">
-            <Text variant="tertiary">
-                {children !== undefined ? children : <Translation id="TR_TROUBLE_SHOOTING_TIPS" />}
-            </Text>
-            <Collapsible.ToggleIcon data-testid="@onboarding/expand-troubleshooting-tips/toggle" />
-        </Row>
+    <Row gap={spacings.xs} cursor="pointer">
+        <Text variant="tertiary" typographyStyle="label">
+            {children !== undefined ? children : <Translation id="TR_TROUBLE_SHOOTING_TIPS" />}
+        </Text>
+        <Collapsible.ToggleIcon
+            data-testid="@onboarding/expand-troubleshooting-tips/toggle"
+            size="medium"
+            iconName="caretCircleDown"
+        />
     </Row>
 );
